labs/05_factories_and_fakes/requirements.txt
# Dependecies for this lab
Flask
Flask-SQLAlchemy
nose
pinocchio
coverage
factory-boy